Output view and model now sync with each other.
No longer have to keep track of every change with the UI or data model backing it. Just call the _sync_with_ui method and everything will sync and update itself. Also added the get_output_path method, which will return the proper path to save the output file, given the path to the input file.